cripto_requestor_app/worker/requirements.txt
pyzmq==25.1.1
pycoingecko==3.1.0
